requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.13.0=pyhd8ed1ab_0
anyio=3.0.1=py37h89c1867_0
argon2-cffi=20.1.0=py37h27cfd23_1
async_generator=1.10=py37h28b3542_0
attrs=21.2.0=pyhd3eb1b0_0
babel=2.9.1=pyh44b312d_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blosc=1.19.0=hd408876_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.5.30=ha878542_0
cairo=1.14.12=h8948797_3
catkin-pkg=0.4.23=pypi_0
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.0=py37h2e261b9_0
chardet=4.0.0=py37h89c1867_1
charls=2.1.0=he6710b0_2
cloudpickle=1.6.0=py_0
cryptography=3.4.7=py37h5d9358c_0
cudatoolkit=11.0.221=h6bb024c_0
cycler=0.10.0=py37_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.4.1=pyhd3eb1b0_0
dataclasses=0.8=pyhc8e2a94_3
dbus=1.13.18=hb2f20db_0
decorator=5.0.7=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
distro=1.6.0=pypi_0
docutils=0.17.1=pypi_0
drawnow=0.72.5=pypi_0
entrypoints=0.3=py37_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
frechetdist=0.6=pypi_0
freetype=2.10.4=h5ab3b9f_0
fribidi=1.0.10=h7b6447c_0
fsspec=0.9.0=pyhd3eb1b0_0
giflib=5.1.4=h14c3975_1
glib=2.63.1=h5a9c865_0
graphite2=1.3.14=h23475e2_0
graphviz=2.40.1=h21bd128_2
grpcio=1.33.2=py37haffed2e_2
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
harfbuzz=1.8.8=hffaf4a1_0
icu=58.2=he6710b0_3
idna=2.10=pyh9f0ad1d_0
imagecodecs=2020.5.30=py37h567f118_1
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
importlib_metadata=3.10.0=hd3eb1b0_0
intel-openmp=2021.2.0=h06a4308_610
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.22.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.0=py37_0
jinja2=3.0.0=pyhd3eb1b0_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py37_7
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_console=6.4.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py37h06a4308_0
jupyter_server=1.7.0=py37h89c1867_1
jupyterlab=3.0.15=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_server=2.5.2=pyhd8ed1ab_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jxrlib=1.1=h7b6447c_2
kiwisolver=1.3.1=py37h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libaec=1.0.4=he6710b0_1
libffi=3.2.1=hf484d3e_1007
libgcc=7.2.0=h69d50b8_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libwebp=1.0.1=h8e7db2f_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
libzopfli=1.0.3=he6710b0_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=pyhd8ed1ab_0
markupsafe=2.0.0=py37h27cfd23_0
matplotlib=3.4.2=pypi_0
missingno=0.4.2=pypi_0
mistune=0.8.4=py37h14c3975_1001
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py37h27cfd23_1
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.1=py37ha9443f7_2
nbclassic=0.2.8=pyhd8ed1ab_0
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py37_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
networkx=2.5=py_0
ninja=1.10.2=hff7bd54_1
notebook=6.3.0=py37h06a4308_0
numpy=1.20.1=py37h93e21f0_0
numpy-base=1.20.1=py37h7d8b39e_0
olefile=0.46=py37_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1l=h7f8727e_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.4=py37h2531618_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py37h06a4308_1
pango=1.42.4=h049681c_0
parso=0.8.2=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.2.0=py37he98fc37_0
pip=21.0.1=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
prometheus_client=0.10.1=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.17.2=py37h295c915_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycparser=2.20=py_2
pydot=1.4.2=pypi_0
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.17.3=py37h7b6447c_0
pysocks=1.7.1=py37h89c1867_3
python=3.7.7=hcf32534_0_cpython
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.17=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
pyzmq=20.0.0=py37h2531618_1
qt=5.9.7=h5867ecd_1
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3deb0d_0
rosdistro=0.8.3=pypi_0
rosinstall=0.7.8=pypi_0
rospkg=1.3.0=pypi_0
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.24.1=py37ha9443f7_0
scipy=1.6.2=py37had2a1c9_1
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py37h06a4308_0
similaritymeasures=0.4.4=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.15.0=py37h06a4308_0
snappy=1.1.8=he6710b0_0
sniffio=1.2.0=py37h89c1867_1
sqlite=3.35.4=hdfb4753_0
tensorboard=1.15.0=py37_0
terminado=0.9.4=py37h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.17=pyhd3eb1b0_1
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchaudio=0.7.2=py37
torchvision=0.8.2=py37_cu110
tornado=6.1=py37h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd8ed1ab_0
vcstools=0.1.42=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py37_1
websocket-client=0.57.0=py37h89c1867_4
werkzeug=2.0.1=pyhd8ed1ab_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37_0
wstool=0.1.17=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
